refactoring to move CaretSelectionPreserver to its own file
